requirements.txt
PyYAML==6.0.2
Jinja2==3.1.4
cpplint==2.0.0
unidecode==1.3.8